api/requirements.txt
# Web framework
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.1.0

# Computer vision (headless version for servers)
opencv-python-headless>=4.5.0
numpy>=1.21.0

# Cube solving
kociemba>=1.2.0

# Image processing
pillow>=8.0.0

# Optional
requests>=2.25.0
